Rename complicated_topics -> complicated_topics_insurance.
Originally, my intention was to keep adding sections to the article for
each new topic I want to discuss. But, I've decided that for my
publishing model, I don't like the idea of fully new ideas being
relegated to the "recently edited" list on the homepage.
These should be separate articles so that the chronological publication
order is more clear and they appear in the RSS feed properly.