The long awaited reveal of the splash text origins.
Due to the extraordinary public demand which I imagine exists, and the
quantity of requests I have received in my email inbox (zero), I have
decided to give in to this crushing pressure and reveal the origins of
the nigh mythical splashtexts. May their je ne sais quoi reach far and
wide.

And a fullscreen button so you can admire them with more attention.