Add link to Wikipedia "Cradle of civilization".
[minor]

I originally chose six thousand years because Sumer / Mesopotamia
saw some of the earliest states in 4000 BC, but I'm changing it to
twelve thousand because the Cradle page says that sedentary and
permanent villages started more like 10,000-12,000 BC, which seems
like a better reference point for the start of trade and economy.
Also because I don't want people to interpret the number six thousand
as a reference to creationism.